tune: MarksArticle move into /Unit, adjust style for ViewSummary.
