+++ public/css/styleUnitCommon.css
@@ -7,8 +7,8 @@
 }
 
 .nodesListSum{
-  font-size: 1.6rem;
-  letter-spacing: 0.17rem;
+  font-size: 1.54rem;
+  letter-spacing: 0.154rem;
   font-weight: 400;
   font-family: 'cwTeXMing';
   color: #FAFAFA;
